feat: add warning for ESP-IDF actions if IDF_TOOLCHAIN is not set to clang
Currently, the ESP-IDF actions, clang-check and clang-html-report,
should be used exclusively with the clang toolchain. Although this is
outlined in the documentation[1], there have been several instances where
these clang actions were used with the gcc toolchain. This update
introduces a clear warning if such a action is used with an incorrect
toolchain.

Closes https://github.com/espressif/clang-tidy-runner/issues/45

[1] https://docs.espressif.com/projects/esp-idf/en/latest/esp32/api-guides/tools/idf-clang-tidy.html

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>